container/2.train_and_eval/2.2.resultAggregation/requirements.txt
numpy
scipy
scikit-learn==1.0.2
google-cloud-storage==2.1.0; python_version > '3.6'